Refactor to follow modular monolith architecture
