+++ src/main/java/enterprise/hoversprite/common/jwt/UserDetailsService.java
@@ -1,6 +1,6 @@
 package enterprise.hoversprite.common.jwt;
 
-import enterprise.hoversprite.modules.user.service.IUserService;
+import enterprise.hoversprite.modules.user.IUserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
@@ -15,7 +15,7 @@ public class UserDetailsService implements org.springframework.security.core.use
     @Override
     public UserDetails loadUserByUsername(String id) throws UsernameNotFoundException {
         try {
-            return userService.getUserById(Long.parseLong(id));
+            return userService.getUserAuthInfoById(Long.parseLong(id));
         } catch (Exception e) {
             throw new UsernameNotFoundException(e.getMessage());
         }
