+++ src/main/java/enterprise/hoversprite/common/types/Location.java
@@ -1,33 +1,15 @@
 package enterprise.hoversprite.common.types;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Embeddable;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Embeddable
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
 public class Location {
     private double latitude;
     private double longitude;
-
-    public Location() {
-    }
-
-    public Location(double latitude, double longitude) {
-        this.latitude = latitude;
-        this.longitude = longitude;
-    }
-
-    public double getLatitude() {
-        return latitude;
-    }
-
-    public void setLatitude(double latitude) {
-        this.latitude = latitude;
-    }
-
-    public double getLongitude() {
-        return longitude;
-    }
-
-    public void setLongitude(double longitude) {
-        this.longitude = longitude;
-    }
 }
