+++ src/main/java/enterprise/hoversprite/modules/auth/dtos/request/SignInRequestDTO.java
@@ -1,4 +1,4 @@
-package enterprise.hoversprite.modules.auth.dtos;
+package enterprise.hoversprite.modules.auth.dtos.request;
 
 import lombok.Data;
 import lombok.NonNull;
