+++ src/main/java/enterprise/hoversprite/modules/user/User.java
@@ -1,30 +1,26 @@
-package enterprise.hoversprite.modules.user.model;
+package enterprise.hoversprite.modules.user;
 
 import enterprise.hoversprite.common.enums.AuthRole;
 import enterprise.hoversprite.common.enums.Expertise;
 import enterprise.hoversprite.common.enums.UserRole;
 import enterprise.hoversprite.common.types.Location;
-import enterprise.hoversprite.modules.user.dtos.UserDTO;
+import enterprise.hoversprite.modules.user.dtos.UserAuthInfoDTO;
+import enterprise.hoversprite.modules.user.dtos.UserInfoDTO;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
 
-import java.util.Collection;
-import java.util.Date;
-import java.util.List;
+import java.time.LocalDate;
 
 @Entity
 @Table(name = "users")
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class User implements UserDetails {
+class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -63,43 +59,17 @@ public class User implements UserDetails {
 
     @CreationTimestamp
     @Column(updatable = false, name = "created_at")
-    private Date createdAt;
+    private LocalDate createdAt;
 
     @UpdateTimestamp
     @Column(name = "updated_at")
-    private Date updatedAt;
+    private LocalDate updatedAt;
 
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return List.of(new SimpleGrantedAuthority(authRole.name()), new SimpleGrantedAuthority(userRole.name()));
+    public UserAuthInfoDTO toUserAuthInfoDto() {
+        return new UserAuthInfoDTO(id, password, userRole, authRole);
     }
 
-    @Override
-    public String getUsername() {
-        return String.valueOf(id);
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return UserDetails.super.isAccountNonExpired();
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return UserDetails.super.isAccountNonLocked();
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return UserDetails.super.isCredentialsNonExpired();
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return UserDetails.super.isEnabled();
-    }
-
-    public UserDTO toDto() {
-        return new UserDTO(id, fullName, phoneNumber, emailAddress, homeAddress, userRole, username, expertise, createdAt, updatedAt);
+    public UserInfoDTO toUserInfoDto() {
+        return new UserInfoDTO(id, fullName, phoneNumber, emailAddress, homeAddress, emailConfirmed, expertise, createdAt, updatedAt);
     }
 }
