+++ src/main/java/enterprise/hoversprite/modules/user/UserRepository.java
@@ -1,13 +1,12 @@
-package enterprise.hoversprite.modules.user.repository;
+package enterprise.hoversprite.modules.user;
 
-import enterprise.hoversprite.modules.user.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByEmailAddress(String email);
 
     Optional<User> findByPhoneNumber(String phoneNumber);
