--- src/main/java/enterprise/hoversprite/modules/user/dtos/DeleteUserResponseDTO.java
@@ -1,10 +1,0 @@
-package enterprise.hoversprite.modules.user.dtos;
-
-import lombok.Data;
-import lombok.NonNull;
-
-@Data
-public class DeleteUserResponseDTO {
-    @NonNull
-    private String message;
-}
