+++ src/main/java/enterprise/hoversprite/modules/user/dtos/response/GetUserResponseDTO.java
@@ -1,5 +1,6 @@
-package enterprise.hoversprite.modules.user.dtos;
+package enterprise.hoversprite.modules.user.dtos.response;
 
+import enterprise.hoversprite.modules.user.dtos.UserInfoDTO;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NonNull;
@@ -10,6 +11,6 @@ public class GetUserResponseDTO {
     @NonNull
     private String message;
 
-    private UserDTO userDTO;
+    private UserInfoDTO userInfo;
 
 }
